fix: refactor helper class for domain tests
