+++ tests/CNAB.Domain.Test/CNAB.Domain.Test.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -20,6 +20,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\CNAB.Domain\CNAB.Domain.csproj" />
+    <ProjectReference Include="..\CNAB.TestHelpers\CNAB.TestHelpers.csproj" />
   </ItemGroup>
 
   <ItemGroup>
